[NO PR INTENTIONAL] Add dummy download component
I'm not creating a pull request because back end is still
researching how to download from S3 so I do not have
specifications for how this component ought to work.

Data flow from parent components is not yet clear but should be
simple. Also, this may need a refactor once the back end starts
implementing downloading.